+++ init_web_v2.go
@@ -360,6 +360,9 @@ func initWebV2Actions(app *App) {
 	m.Get("/v3/data_owner_update_requests", handlers.GetDataOwnerUpdateRequests)
 	m.Get("/v3/data_owner_update_requests/{id}", handlers.GetDataOwnerUpdateRequests)
 
+	m.Get("/v3/liquidity-pools", handlers.GetLiquidityPoolList)
+	m.Get("/v3/liquidity-pools/{id}", handlers.GetLiquidityPool)
+
 	cop := app.config.Copus()
 	if err := cop.RegisterChi(m); err != nil {
 		panic(errors.Wrap(err, "failed to register service"))